Restructure plugins files
