+++ RecoHGCal/TICL/plugins/ClusterFilterBase.h
