+++ RecoHGCal/TICL/plugins/ClusterFilterFactory.h
@@ -3,7 +3,7 @@
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/PluginManager/interface/PluginFactory.h"
-#include "RecoHGCal/TICL/interface/ClusterFilterBase.h"
+#include "ClusterFilterBase.h"
 
 typedef edmplugin::PluginFactory<ticl::ClusterFilterBase*(const edm::ParameterSet&)> ClusterFilterFactory;
